Fix harmless signed/unsigned comparison warnings in Firebird code.
At least MSVC warns about comparing unsigned std::string::size() result with a
signed value, fix the warning by simply converting the former to signed int.

+++ src/backends/firebird/common.h
@@ -161,7 +161,7 @@ std::string format_decimal(const void *sqldata, int sqlscale)
     std::string r = out.str();
     if (sqlscale < 0)
     {
-        if (r.size() - (x < 0) <= -sqlscale)
+        if (static_cast<int>(r.size()) - (x < 0) <= -sqlscale)
         {
             r = std::string(size_t(x < 0), '-') +
                 std::string(-sqlscale - (r.size() - (x < 0)) + 1, '0') +
